+++ CHANGES.md
@@ -1,3 +1,4 @@
 **[Infinite Elytra Parkour](https://www.spigotmc.org/resources/115322/) | [Infinite Parkour+](https://www.spigotmc.org/resources/105019/)**
 
 - Added PAPI placeholder for the current mode
+- Fix leaves disappearing in parkour